requirements.txt
streamlit
pandas
plotly
Pillow
db-sqlite3
python-dotenv
web3
eth-account
streamlit-wallet-connect
matplotlib
